Some updates from hotaru

+++ teiler/install.sh
@@ -1,9 +1,12 @@
-#!/usr/bin/zsh
+#!/usr/bin/env zsh
 
-sudo apt install maim ffmpeg xclip xininfo copyq
-mkdir ~/scratch
-git clone https://github.com/carnager/teiler.git ~/scratch/teiler
-pushd ~/scratch/teiler
-sudo make install
-popd
-rm -rf ~/scratch/teiler
+if [ "$(uname -s)" != "Darwin" ]
+then
+	sudo apt install maim ffmpeg xclip xininfo copyq
+	mkdir ~/scratch
+	git clone https://github.com/carnager/teiler.git ~/scratch/teiler
+	pushd ~/scratch/teiler
+	sudo make install
+	popd
+	rm -rf ~/scratch/teiler
+fi
